주석 정리, import문 정리, 사용하지 않는 코드 정리, 회원탈퇴 변수 수정
