+++ tomyanimal/src/Animals/Checklist/CardItem.js
@@ -1,5 +1,5 @@
-import React, { useContext, useState } from 'react'
-import { Link, useNavigate } from 'react-router-dom'
+import React, { useContext } from 'react'
+import { useNavigate } from 'react-router-dom'
 import { ChecklistContext } from './CheckList'
 import './CardItem.css'
 
@@ -7,13 +7,11 @@ const CardItem = ( props, {id, text, src, label} ) => {
     const navigate = useNavigate();
 
     const checklist = useContext(ChecklistContext);
-    //console.log(checklist);
 
   return (
     <>
         {checklist.map((it) => 
-        <li className='cards__item'>
-            {/* <Link className='cards__item__link' to={props.path}> */}
+            <li className='cards__item'>
                 <div 
                     className='cards__item__pic__wrap' 
                     data-category={it.label}
@@ -28,13 +26,11 @@ const CardItem = ( props, {id, text, src, label} ) => {
                     />
                 </div>
                 <div className='cards__item__info'>
-                    {/* <h4 className='cards__item__text'>{props.text}</h4> */}
                     <h4 className='cards__item__text'>{it.text}</h4>
                 </div>
 
-            {/* </Link> */}
-        </li>
-            )}
+            </li>
+        )}
     </>
   )
 }
