+++ tomyanimal/src/Animals/Checklist/CheckList.js
@@ -1,9 +1,6 @@
-import React, { useState, createContext } from 'react'
-import { Link } from 'react-router-dom';
+import React from 'react'
 import CardItem from './CardItem';
-import ControlMenu from '../../Pages/ControlMenu';
 //import img from './img/imageex.png'
-import Walk from './Walk';
 
 export const ChecklistContext = React.createContext();
 
@@ -54,7 +51,6 @@ const checklistTab = [
 ]
 
 const CheckList = () => {
-  //console.log(checklistTab);
 
   return (
     <ChecklistContext.Provider value={checklistTab}>
@@ -76,89 +72,7 @@ const CheckList = () => {
               <div className='checklist__content'>
                 <CardItem />
               </div>
-
-              {/* <div className='checklist__content'>
-                <CardItem 
-                  text = "산책"
-                  src = "/img/imageex.png"
-                  label = 'Walking'
-                  path='/animal/walk'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "사료"
-                  src = "/"
-                  label = 'Walking'
-                  path='/'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "간식"
-                  src = "/"
-                  label = 'Walking'
-                  path='/animal'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "미용"
-                  src = "/"
-                  label = 'Walking'
-                  path='/animal'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "진료"
-                  src = "/"
-                  label = 'Walking'
-                  path='/animal'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "목욕"
-                  src = "/"
-                  label = 'Walking'
-                  path='/animal'
-                />
-              </div>
-
-              <div className='checklist__content'>
-                <CardItem 
-                  text = "대소변"
-                  src = "/"
-                  label = 'Walking'
-                  path='/animal'
-                />
-              </div> */}
-
-              {/*
-              <div className='checklist__content'>
-                {checklistTab.map((it) => 
-                <CardItem
-                  key={it.id}
-                  {...it}
-                />
-                )}
-              </div>
-                
-              
-              <div className='checklist__content'>
-                <Link to="/animal" >
-                  산책
-                  <div className='checklist__img'><img src={img} /></div>
-                </Link>
-              </div>
-              */}
-              
+            
             </div>
           </section>
         </section>
