+++ tomyanimal/src/Users/components/ReadOnlyRow.js
@@ -4,10 +4,6 @@ const ReadOnlyRow = ( {userPhone, userName, handleEditClick}) => {
   return (
   <>
     <tr>
-      {/* <td>{user.name}</td>
-      <td>{user.contact}</td> */}
-      {/* <td>{user.result.data['userName']}</td>
-      <td>{user.result.data['userPhoneNumber']}</td> */}
       <td>{userName}</td>
       <td>{userPhone}</td>
