--- tomyanimal/src/Users/data.js
@@ -1,12 +1,0 @@
-export const userData = [
-    {
-        "id": 1, 
-        "name": "Yeji Kim",
-        "contact": "01011112222",
-      },
-      // {
-      //   "id": 2, 
-      //   "name": "Junseok Lee",
-      //   "contact": "01022342234",
-      // },
-]
